refactor: Moving to root of xero/lib.
