+++ ledgers/xero/lib/resource.rb
