+++ ledgers/xero/spec/spec_helper.rb
@@ -1,5 +1,5 @@
 require "bundler/setup"
-require "ledger_sync/ledgers/xero"
+require "xero"
 
 RSpec.configure do |config|
   # Enable flags like --only-failures and --next-failure
